Test refactoring. All the load of the test was made on TestRoverController, which made sense when the logic was all there, but since It got refactored and splitted in different classes, the tests shouldn't really be there.
I moved the majority of the tests and there is now proper unit testing of every single class, instead of going through every test using the controller. It should make for a more clean and correct, as we are now actually doing unitary testing.

It also allowed me to remove some aditional methods in rovercontroller which needed to be there due to the way we were testing. We no longer need them.